removing ws-worker changes
